Use the new ClassifierTrainer inheritance instead of repeating the train_loop
